refactor: utils::net is one file
